working on cost tracking with spans
